feat(): add asset history API
